+++ src/main/kotlin/com/dcd/server/core/domain/application/spi/ImageVersionPort.kt
@@ -1,0 +1,5 @@
+package com.dcd.server.core.domain.application.spi
+
+interface ImageVersionPort {
+    fun fetchAllVersions(imageName: String): List<String>
+}
